[ feat ] 19 : 구매 프로젝트 컨트롤러 위치 변경
